Remove ALLOW_TELEPORT config option, add ALLOW_EXIT_BOUNDARY
A config option to disable teleportation entirely is beyond the scope of
the framework, and by defaulting to true, may confuse developers by
cancelling their teleport calls. An ALLOW_EXIT_BOUNDARY config option
provides control to minigames that was previously missing by giving a
toggle for a certain type of internal behavior.