Add framework for better handling of orphaned objects
This commit adds the @Orphanable annotation, meant to denote a class
which may be orphaned. It also adds an OrphanedObjectExcepotion, thrown
by methods of an object annotated with @Orphanable if it is orphaned
(this replaces the IllegalStateException thrown by previous snapshots in
this circumstance). Finally, Arenas, Challengers, Teams, and Rounds are
all now orphanable (previously only Challengers could be marked as
orphaned).